pyproject.toml
[project]
name = "fasttts"
version = "0.1.0"
description = "A Text-to-Speech API server supporting multiple TTS engines"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "fastapi[all]>=0.128.0",
    "realtimetts[all]>=0.5.7",
    "pydub>=0.25.0",
    "websockets>=12.0",
]

[project.scripts]
fasttts = "fasttts.app:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
fasttts = ["py.typed"]
